Refactor main page and fix users search validation #42
