+++ client/src/pages/main.page.tsx
@@ -1,5 +1,6 @@
 import React, { ReactElement, useState, useEffect } from "react"
 import styled from "styled-components"
+import { useSockets } from "../effects/use-sockets"
 
 import { UsersList } from "../components/users/users-list"
 import { MessagesList } from "../components/messages/messages-list"
@@ -7,7 +8,6 @@ import { SignOut } from "../components/main/sign-out"
 
 import { User, Message } from "../utils/interfaces"
 import { theme } from "../style-guide/theme"
-import { socketUrl } from "../utils/api-utils"
 
 interface Props {
 	user: User
@@ -52,11 +52,11 @@ export const MainPage = ({
 	const [newMessage, setNewMessage] = useState<Message>()
 	const [onlineFriends, setOnlineFriends] = useState<number[]>()
 
-	const [connectionsNumber, setconnectionsNumber] = useState(1)
-
 	const [originalFriends, setOriginalFriends] = useState<User[]>(
 		user.friends || [],
 	)
+
+	const { data } = useSockets()
 
 	const updateUsersList = (users?: User[] | null): void => {
 		if (users) {
@@ -75,59 +75,41 @@ export const MainPage = ({
 	}
 
 	useEffect((): void => {
-		if (!connectionsNumber) {
+		if (!data) {
 			return
 		}
 
-		const ws = new WebSocket(socketUrl)
+		if (data.newMessage) {
+			const newMessage = data.newMessage
+			setNewMessage(newMessage)
 
-		ws.onmessage = (event): void => {
-			const data = JSON.parse(event.data)
-
-			if (data.newMessage) {
-				const newMessage = data.newMessage
-				setNewMessage(newMessage)
-
-				if (
-					!friends ||
-					!friends.find(
-						(friend): boolean => friend.id === newMessage.from,
-					)
-				) {
-					setFriends([
-						...(friends || []),
-						{
-							id: newMessage.from,
-							name: data.fromName,
-							email: "",
-							isOnline: true,
-						},
-					])
-				}
-			}
-
-			if (data.online) {
-				setOnlineFriends(
-					data.online.map(
-						(onlineUser: { userId: number }): number =>
-							onlineUser.userId,
-					),
+			if (
+				!friends ||
+				!friends.find(
+					(friend): boolean => friend.id === newMessage.from,
 				)
+			) {
+				setFriends([
+					...(friends || []),
+					{
+						id: newMessage.from,
+						name: data.fromName,
+						email: "",
+						isOnline: true,
+					} as User,
+				])
 			}
 		}
 
-		ws.onclose = (): void => {
-			setTimeout(
-				(): void => setconnectionsNumber(connectionsNumber + 1),
-				100 * connectionsNumber,
+		if (data.online) {
+			setOnlineFriends(
+				data.online.map(
+					(onlineUser: { userId: number }): number => onlineUser.userId,
+				),
 			)
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [connectionsNumber])
-
-	useEffect((): void => {
-		setconnectionsNumber(0)
-	}, [])
+	}, [data])
 
 	const addNewFriend = (userId: number): void => {
 		if (
